feat(reset): return an overlay installation to native mode
The workspace row survives the sweep because it carries the organization, so
x_sor_mode and x_incumbent survived with it -- while everything overlay mode
depends on was swept out from under them. A reset left the installation
claiming to read from an incumbent it no longer had a connection to, with
every read dispatching at an empty mirror: worse than not resetting, because
it looks like it works.

The flip is overlay.RevertToNative, exported from the module that owns those
fork-owned columns rather than spelled a second time in compose, and
Disconnect's teardown now takes the same function. Both columns move in one
statement because the x_overlay_iff_incumbent CHECK admits no intermediate
state, and the predicate makes it idempotent so a native install is not
written to. It is deliberately NOT the governed Disconnect teardown: the
connection and mirror rows are already gone, the reset has its own audit row,
and incumbent.disconnected would be staged into an outbox this reset just
drained.

datareset.go crossed the 500-line ceiling, so the Postgres sweep primitives
move to datasweep.go by concept.
